requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
cffi==1.16.0
discord.py==2.3.2
Django==5.0.6
django-cors-headers==4.3.1
frozenlist==1.4.1
idna==3.7
lxml==5.2.2
multidict==6.0.5
pycparser==2.22
PyNaCl==1.5.0
python-docx==1.1.2
pytz==2024.1
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
yarl==1.9.4
